WIP refactoring to make it easier to use sentry/splunk
