+++ server/app.js
@@ -52,25 +52,7 @@ app.use(setAllocationID);
 // Set the A/B allocation as a response header
 app.use(setAllocationHeader);
 
-app.get('/*', function(req, res) {
-
-	// Metrics: Who's asking for the allocation?
-	// FIXME - var interrogator = req.get('ft-interrogator') || 'unknown';
-	// metrics.count('interrogator.'+interrogator, 1);
-	if (req.get('ft-session-token') || req.get('ft-allocation-id')) {
-		res
-			.set('Cache-Control', 'max-age=10800, public, stale-while-revalidate=10800, stale-if-error=86400')
-			.set('Outbound-Cache-Control', 'private, max-age=0, no-cache')
-			.set('Vary', 'ft-allocation-id, ft-session-token');
-	} else {
-		res
-			.set('Cache-Control', 'private, max-age=0, no-cache');
-	}
-
-	res
-		.status(200)
-		.send('OK');
-});
+app.get('/*', require('./controllers/allocate'));
 
 module.exports.listen = app.listen(process.env.PORT, function() {
 	console.log('Listening on port', process.env.PORT);
