+++ server/lib/allocate-tests.js
@@ -7,16 +7,6 @@ const seedrandom = require('seedrandom');
 // but evenly allocate users into an A/B segment, per test.
 module.exports = function(tests, user) {
 
-	if (!user || !user.uuid) {
-		metrics.count('allocation.failed.uuid', 1);
-		return false;
-	}
-
-	if (!tests || tests.flagsWithABTests.length === 0) {
-		metrics.count('allocation.failed.tests', 1);
-		return false;
-	}
-
 	metrics.count('allocation.count', 1);
 
 	const allocatedTests = user.isSubscriber ? tests.subscriberTests : tests.anonymousTests;
